# backend_reservation_optimizer/src/api-spec.yaml
openapi: 3.0.0
info:
  title: Zoom Meeting Management API
  version: 1.0.0
paths:
...
  /users/login:
    post:
      summary: User login
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  format: username
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Authenticated JWT for the user
        '401':
          description: Unauthorized
...
  /users/{userId}:
    put:
      summary: Update a user
      tags:
        - Users
      parameters:
        - name: userId
          in: path
          required: true
          description: ID of the user to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated successfully
        '404':
          description: User not found
...
components:
  schemas:
    User:
      type: object
      properties:
        username:
          type: string
        fistname:
          type: string
        lastname:
          type: string
        passwordHash:
          type: string
        role:
          type: string
          enum:
            - admin
            - standard_user
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...